Split wwatch3 forecast run into nowcast, then forecast.

+++ nowcast/workers/run_ww3.py
@@ -12,16 +12,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Salish Sea WaveWatch3 forecast worker that prepares the temporary run
-directory and bash run script for a prelim-forecast or forecast run on the 
-ONC cloud, and launches the run.
+"""Salish Sea WaveWatch3 nowcast/forecast worker that prepares the temporary
+run directory and bash run script for a prelim-forecast, nowcast or forecast
+run on the ONC cloud, and launches the run.
 """
 import logging
 import os
 from pathlib import Path
 import shlex
 import subprocess
-import uuid
 
 import arrow
 from nemo_nowcast import NowcastWorker
@@ -44,11 +43,12 @@ def main():
     worker.cli.add_argument("host_name", help="Name of the host to execute the run on")
     worker.cli.add_argument(
         "run_type",
-        choices={"forecast2", "forecast"},
+        choices={"forecast2", "nowcast", "forecast"},
         help="""
         Type of run to execute:
         'forecast2' means preliminary forecast run (after NEMO forecast2 run),
-        'forecast' means updated forecast run (after NEMO forecast run)
+        'nowcast' means nowcast run (after NEMO forecast run)
+        'forecast' means updated forecast run (after WaveWatch3 forecast run)
         """,
     )
     worker.cli.add_date_option(
@@ -139,20 +139,23 @@ def _build_tmp_run_dir(run_date, run_type, config):
     :rtype: :py:class:`pathlib.Path`
     """
     run_prep_path = Path(config["wave forecasts"]["run prep dir"])
-    run_dir_path = _make_run_dir(run_prep_path)
+    run_dir_path = _make_run_dir(run_type, run_prep_path)
     _create_symlinks(run_date, run_type, run_prep_path, run_dir_path, config)
     _write_ww3_input_files(run_date, run_type, run_dir_path)
     return run_dir_path
 
 
-def _make_run_dir(run_prep_dir):
+def _make_run_dir(run_type, run_prep_dir):
     """
+    :param str run_type:
     :param :py:class:`pathlib.Path` run_prep_dir:
     
     :return: Temporary run directory
     :rtype: :py:class:`pathlib.Path`
     """
-    run_dir_path = run_prep_dir / str(uuid.uuid1())
+    run_dir_path = (
+        run_prep_dir / f'{run_type}_{arrow.now().format("YYYY-MM-DDTHHmmss.SSSSSSZ")}'
+    )
     run_dir_path.mkdir(mode=0o775)
     return run_dir_path
 
@@ -167,8 +170,9 @@ def _create_symlinks(run_date, run_type, run_prep_path, run_dir_path, config):
     """
     for target in ("mod_def.ww3", "wind", "current"):
         (run_dir_path / target).symlink_to(run_prep_path / target)
-    results_path = Path(config["wave forecasts"]["results"][run_type])
-    ddmmmyy = run_date.shift(days=-1).format("DDMMMYY").lower()
+    results_path = Path(config["wave forecasts"]["results"]["nowcast"])
+    restart_date = run_date.shift(days=-1) if run_type == "nowcast" else run_date
+    ddmmmyy = restart_date.format("DDMMMYY").lower()
     prev_run_path = results_path / ddmmmyy
     (run_dir_path / "restart.ww3").symlink_to(prev_run_path / "restart001.ww3")
 
@@ -259,10 +263,18 @@ def _ww3_shel_contents(run_date, run_type):
     :return: ww3_shel.inp file contents
     :rtype: str 
     """
-    start_date = run_date.format("YYYYMMDD")
-    restart_date = run_date.shift(days=+1).format("YYYYMMDD")
-    end_date = run_date.shift(days=+2).format("YYYYMMDD")
-    end_time = "120000" if run_type == "forecast" else "060000"
+    start_date = (
+        run_date.format("YYYYMMDD")
+        if run_type == "nowcast"
+        else run_date.shift(days=+1).format("YYYYMMDD")
+    )
+    end_date = (
+        run_date.shift(days=+1).format("YYYYMMDD")
+        if run_type == "nowcast"
+        else run_date.shift(days=+2).format("YYYYMMDD")
+    )
+    end_times = {"forecast2": "060000", "nowcast": "000000", "forecast": "120000"}
+    end_time = end_times[run_type]
     contents = f"""$ WAVEWATCH III shell input file
 $
 $ Forcing/inputs to use
@@ -302,7 +314,7 @@ $ Start time (YYYYMMDD HHmmss), Interval (s), End time (YYYYMMDD HHmmss)
 $
 $ Restart files
 $ Start time (YYYYMMDD HHmmss), Interval (s), End time (YYYYMMDD HHmmss)
-  {restart_date} 000000 3600 {restart_date} 000000
+  {end_date} 000000 3600 {end_date} 000000
 $
 $ Boundary data (required placeholder for unused feature)
 $ Start time (YYYYMMDD HHmmss), Interval (s), End time (YYYYMMDD HHmmss)
@@ -327,11 +339,18 @@ def _ww3_ounf_contents(run_date, run_type):
     :return: ww3_ounf.inp file contents
     :rtype: str 
     """
-    start_date = run_date.format("YYYYMMDD")
+    start_date = (
+        run_date.format("YYYYMMDD")
+        if run_type == "nowcast"
+        else run_date.shift(days=+1).format("YYYYMMDD")
+    )
+    run_hours = {"nowcast": 24, "forecast": 36, "forecast2": 30}
+    output_interval = 1800  # seconds
+    output_count = int(run_hours[run_type] * 60 * 60 / output_interval)
     contents = f"""$ WAVEWATCH III NETCDF Grid output post-processing
 $
 $ First output time (YYYYMMDD HHmmss), output increment (s), number of output times
-  {start_date} 000000 1800 120
+  {start_date} 000000 {output_interval} {output_count}
 $
 $ Fields
   N  by name
@@ -366,11 +385,18 @@ def _ww3_ounp_contents(run_date, run_type):
     :return: ww3_ounp.inp file contents
     :rtype: str 
     """
-    start_date = run_date.format("YYYYMMDD")
+    start_date = (
+        run_date.format("YYYYMMDD")
+        if run_type == "nowcast"
+        else run_date.shift(days=+1).format("YYYYMMDD")
+    )
+    run_hours = {"nowcast": 24, "forecast": 36, "forecast2": 30}
+    output_interval = 600  # seconds
+    output_count = int(run_hours[run_type] * 60 * 60 / output_interval)
     contents = f"""$ WAVEWATCH III NETCDF Point output post-processing
 $
 $ First output time (YYYYMMDD HHmmss), output increment (s), number of output times
-  {start_date} 000000 600 360
+  {start_date} 000000 {output_interval} {output_count}
 $
 $ All points defined in ww3_shel.inp
   -1
@@ -392,7 +418,6 @@ $ WMO standard output
   6
 """
 
-    start_date = run_date.format("YYYYMMDD")
     return contents
 
 
@@ -424,8 +449,8 @@ def _build_run_script(run_date, run_type, run_dir_path, results_path, config):
                     run_date, run_type, run_dir_path, results_path, config
                 ),
                 prepare=_prepare(),
-                execute=_execute(run_date),
-                netcdf_output=_netcdf_output(run_date),
+                execute=_execute(run_type, run_date),
+                netcdf_output=_netcdf_output(run_date, run_type),
                 cleanup=_cleanup(),
             ),
         )
@@ -490,8 +515,9 @@ def _prepare():
     return preparations
 
 
-def _execute(run_date):
+def _execute(run_type, run_date):
     """
+    :param str run_type:
     :param :py:class:`arrow.Arrow` run_date:
 
     :return: Execution section of wwatch3 run set-up and execution script
@@ -504,37 +530,43 @@ def _execute(run_date):
         "  >>${RESULTS_DIR}/stdout 2>>${RESULTS_DIR}/stderr && \\\n"
         "mv log.ww3 ww3_shel.log && \\\n"
         "rm current.ww3 wind.ww3 && \\\n"
-        f"rm current/SalishSea_1h_{start_date}_{start_date}_grid_[UV].nc && \\\n"
-        f"rm current/SoG_current_{start_date}.nc && \\\n"
-        f"rm wind/SoG_wind_{start_date}.nc\n"
-        'echo "Ended run at $(date)" >>${RESULTS_DIR}/stdout\n'
     )
+    if run_type != "nowcast":
+        execution += (
+            f"rm current/SoG_current_{start_date}.nc && \\\n"
+            f"rm wind/SoG_wind_{start_date}.nc\n"
+        )
+    execution += 'echo "Ended run at $(date)" >>${RESULTS_DIR}/stdout\n'
     return execution
 
 
-def _netcdf_output(run_date):
+def _netcdf_output(run_date, run_type):
     """
-    :param :py:class:`arrow.Arrow` run_date: 
+    :param str run_type:
+    :param :py:class:`arrow.Arrow` run_date:
     
     :return: netCDF output files section of wwatch3 run set-up and execution 
              script
     :rtype: str 
     """
-    start_yyyymmdd = run_date.format("YYYYMMDD")
-    end_yyyymmdd = run_date.shift(days=+2).format("YYYYMMDD")
+    start_date = run_date if run_type == "nowcast" else run_date.shift(days=+1)
+    start_yyyymmdd = start_date.format("YYYYMMDD")
+    end_date = run_date if run_type == "nowcast" else run_date.shift(days=+2)
+    end_yyyymmdd = end_date.format("YYYYMMDD")
     fields_files = " ".join(
         f'SoG_ww3_fields_{day.format("YYYYMMDD")}.nc'
-        for day in arrow.Arrow.range("day", run_date, run_date.shift(days=+2))
+        for day in arrow.Arrow.range("day", start_date, end_date)
     )
     points_files = " ".join(
         f'SoG_ww3_points_{day.format("YYYYMMDD")}_tab.nc'
-        for day in arrow.Arrow.range("day", run_date, run_date.shift(days=+2))
+        for day in arrow.Arrow.range("day", start_date, end_date)
     )
+    nc_operator = "ncks" if run_type == "nowcast" else "ncrcat"
     output_to_netcdf = (
         'echo "Starting netCDF4 fields output at $(date)" '
         ">>${RESULTS_DIR}/stdout\n"
         "${WW3_EXE}/ww3_ounf >>${RESULTS_DIR}/stdout 2>>${RESULTS_DIR}/stderr && \\\n"
-        f"ncrcat -4 -L4 -o SoG_ww3_fields_{start_yyyymmdd}_{end_yyyymmdd}.nc \\\n"
+        f"{nc_operator} -4 -L4 -o SoG_ww3_fields_{start_yyyymmdd}_{end_yyyymmdd}.nc \\\n"
         f"  {fields_files} \\\n"
         "  >>${RESULTS_DIR}/stdout 2>>${RESULTS_DIR}/stderr && \\\n"
         f"rm {fields_files} && \\\n"
@@ -545,7 +577,7 @@ def _netcdf_output(run_date):
         'echo "Starting netCDF4 points output at $(date)" '
         ">>${RESULTS_DIR}/stdout\n"
         "${WW3_EXE}/ww3_ounp >>${RESULTS_DIR}/stdout 2>>${RESULTS_DIR}/stderr && \\\n"
-        f"ncrcat -4 -L4 -o SoG_ww3_points_{start_yyyymmdd}_{end_yyyymmdd}.nc \\\n"
+        f"{nc_operator} -4 -L4 -o SoG_ww3_points_{start_yyyymmdd}_{end_yyyymmdd}.nc \\\n"
         f"  {points_files} \\\n"
         "  >>${RESULTS_DIR}/stdout 2>>${RESULTS_DIR}/stderr && \\\n"
         f"rm {points_files} && \\\n"
